tools/life-admin/requirements.txt
# Life Admin Tools — Python dependencies
anthropic>=0.40.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
